pyproject.toml
[project]
name = "training-eval-overlap"
version = "0.1.0"
description = "Quantifying overlap between training data (Carolina corpus) and evaluation benchmarks (PoetaV2) for Portuguese LLMs."
authors = [
    {name = "Gustavo Hochgraf", email = "gusthoch@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "pandas>=2.2",
    "datasets>=3.0",
    "pyarrow>=15.0",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "ipykernel>=6.29",
    "matplotlib>=3.9",
    "seaborn>=0.13",
]
search = [
    "rank-bm25>=0.2",
    "datasketch>=1.6",       # MinHash LSH for near-duplicate detection
    "simhash>=2.1",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
